Move PluginManager out of api folder
