+++ common/lib/xmodule/xmodule/tabs.py
@@ -8,7 +8,7 @@ from django.core.files.storage import get_storage_class
 from six import text_type
 from xblock.fields import List
 
-from openedx.core.lib.api.plugins import PluginError
+from openedx.core.lib.plugins import PluginError
 
 log = logging.getLogger("edx.courseware")
