+++ openedx/core/lib/tests/test_course_tab_api.py
@@ -5,7 +5,7 @@ Tests for the plugin API
 from django.test import TestCase
 from nose.plugins.attrib import attr
 
-from openedx.core.lib.api.plugins import PluginError
+from openedx.core.lib.plugins import PluginError
 from openedx.core.lib.course_tabs import CourseTabPluginManager
